--- a/.changes/dynamic-dispatch-async-commands.md
+++ b/.changes/dynamic-dispatch-async-commands.md
@@ -1,0 +1,5 @@
+---
+tauri: "patch:perf"
+---
+
+Use dynamic dispatch for async commands in dev, this should speed up the compilation time by quite a bit, and significantly reduces the incremental compilation time