--- a/crates/tauri/src/ipc/mod.rs
+++ b/crates/tauri/src/ipc/mod.rs
@@ -345,14 +345,31 @@ impl<R: Runtime> InvokeResolver<R> {
   where
     F: Future<Output = Result<InvokeResponseBody, InvokeError>> + Send + 'static,
   {
-    self.respond_async_serialized_dyn(Box::pin(task))
+    // Dynamic dispatch the call in dev for a faster compile time
+    // TODO: Revisit this and see if we can do this for the release build as well if the performace hit is not a problem
+    #[cfg(debug_assertions)]
+    {
+      self.respond_async_serialized_dyn(Box::pin(task))
+    }
+    #[cfg(not(debug_assertions))]
+    {
+      self.respond_async_serialized_inner(task)
+    }
   }
 
-  /// Reply to the invoke promise with an async task which is already serialized.
+  /// Dynamic dispatch the [`Self::respond_async_serialized`] call
   fn respond_async_serialized_dyn(
     self,
     task: Pin<Box<dyn Future<Output = Result<InvokeResponseBody, InvokeError>> + Send + 'static>>,
   ) {
+    self.respond_async_serialized_inner(task)
+  }
+
+  /// Reply to the invoke promise with an async task which is already serialized.
+  fn respond_async_serialized_inner<F>(self, task: F)
+  where
+    F: Future<Output = Result<InvokeResponseBody, InvokeError>> + Send + 'static,
+  {
     crate::async_runtime::spawn(async move {
       let response = match task.await {
         Ok(ok) => InvokeResponse::Ok(ok),