<!--
Before submitting a PR, please read https://github.com/tauri-apps/tauri/blob/dev/.github/CONTRIBUTING.md#pull-request-guidelines

1. Give the PR a descriptive title.

  Examples of good title:
    - fix(windows): fix race condition in event loop
    - docs: update example for `App::show`
    - feat: add `Window::set_fullscreen`

  Examples of bad title:
    - fix #7123
    - update docs
    - fix bugs

2. If there is a related issue, reference it in the PR text, e.g. closes #123.
3. If this change requires a new version, then add a change file in `.changes` directory with the appropriate bump, see https://github.com/tauri-apps/tauri/blob/dev/.changes/README.md
4. Ensure that all your commits are signed https://docs.github.com/en/authentication/managing-commit-signature-verification/signing-commits
5. Ensure `cargo test` and `cargo clippy` passes.
6. Propose your changes as a draft PR if your work is still in progress.
-->

Reference: #13382

From some testing, this could speed up the compile time by ~10s (80s -> 70s) on a full debug build on my machine with a minimal tauri app setup, and reduces the incremental compilation time from ~10s to ~4s (By adding and removing a new line inside an async command)

I'm not entirely sure the performance implications of doing dynamic dispatch for those commands though, maybe we could enable this for the debug build and still use the old static dispatch one for the release build? (did this for dev only right now, we can revisit this in the future)

Some `cargo llvm-lines` outputs

<details><summary>Before</summary>

```
  Lines                  Copies               Function name
  -----                  ------               -------------
  1231833                66222                (TOTAL)
    86143 (7.0%,  7.0%)   2497 (3.8%,  3.8%)  std::panic::catch_unwind
    42500 (3.5%, 10.4%)   2500 (3.8%,  7.5%)  std::panicking::try::do_catch
    34162 (2.8%, 13.2%)   2500 (3.8%, 11.3%)  std::panicking::try::do_call
    33280 (2.7%, 15.9%)    416 (0.6%, 11.9%)  tokio::runtime::task::core::Cell<T,S>::new
    33280 (2.7%, 18.6%)    416 (0.6%, 12.6%)  tokio::runtime::task::harness::poll_future
    29952 (2.4%, 21.1%)    832 (1.3%, 13.8%)  tokio::runtime::task::harness::Harness<T,S>::complete::{{closure}}
    24472 (2.0%, 23.0%)    648 (1.0%, 14.8%)  <alloc::boxed::Box<T,A> as core::ops::drop::Drop>::drop
    22464 (1.8%, 24.9%)    416 (0.6%, 15.4%)  tokio::runtime::task::harness::Harness<T,S>::poll_inner
```

</details> 

<details><summary>After</summary>

```
  Lines                 Copies               Function name
  -----                 ------               -------------
  399528                17715                (TOTAL)
   19985 (5.0%,  5.0%)      2 (0.0%,  0.0%)  tauri_test::main::{{closure}}
   14321 (3.6%,  8.6%)    148 (0.8%,  0.8%)  std::sync::mpmc::context::Context::with::{{closure}}
   14268 (3.6%, 12.2%)     73 (0.4%,  1.3%)  tauri::window::plugin::init::{{closure}}::{{closure}}::{{closure}}
   10552 (2.6%, 14.8%)    203 (1.1%,  2.4%)  <futures_util::future::future::map::Map<Fut,F> as core::future::future::Future>::poll
    8829 (2.2%, 17.0%)    203 (1.1%,  3.6%)  futures_util::future::future::map::_::<impl futures_util::future::future::map::Map<Fut,F>>::project_replace
    5700 (1.4%, 18.4%)     21 (0.1%,  3.7%)  std::sync::mpmc::list::Channel<T>::start_recv
    5581 (1.4%, 19.8%)    203 (1.1%,  4.8%)  futures_util::future::future::map::_::<impl futures_util::future::future::map::Map<Fut,F>>::project
    5292 (1.3%, 21.2%)     63 (0.4%,  5.2%)  std::sync::mpmc::counter::Receiver<C>::release
    4597 (1.2%, 22.3%)     21 (0.1%,  5.3%)  std::sync::mpmc::list::Channel<T>::discard_all_messages
    4350 (1.1%, 23.4%)    108 (0.6%,  5.9%)  <alloc::boxed::Box<T,A> as core::ops::drop::Drop>::drop
```
</details> 

Note: the zip file contains the `cargo build --timings` HTML output, and because GitHub doesn't seem to like uploading HTML files directly, I zipped them in

[cargo-build-timings.zip](https://github.com/user-attachments/files/20273733/cargo-build-timings.zip)