@Legend-Master 2025-05-29
I'll merge this for now since I'm preparing a follow up PR, would be nice to have some comparison data on mac and linux though 🙃